refactor(assistants): extract DetectedFile, UploadsDetector and shared upload utils (EPMCDME-10894)
